kolektif_akil_platform/screen_translator/requirements.txt
pyautogui==0.9.54
pytesseract==0.3.10
googletrans==4.0.0rc1
Pillow==10.0.1